refactor(event-util) : 패키지 변경
