+++ src/main/java/christmas/model/WeekdayEventStrategy.java
@@ -4,6 +4,7 @@ import static christmas.util.Constants.MONTH;
 import static christmas.util.Constants.YEAR;
 import static christmas.util.Constants.ZERO;
 
+import christmas.util.EventUtils;
 import christmas.util.enums.EventType;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
